Generated the 2x2 panel figure and update hard-coded file paths to unix format.
In the bottom of calc_rot_curve(), I added the code to create a two by two paneled figure containing the following:

unmasked H-alpha velocity
masked H-alpha velocity
H-alpha velocity processed in the algorithm
average and stellar rotation curves plotted side by side.

This figure was created so that we can examine the mask placed on the data and how it affects the output of our algorithm.

In addition, I updated hard-coded strings from window's format (\\) to unix format (/).  Python can mitigate this locally on my windows machine so long as any existing cache file is deleted so that the script may 'reset.'