Updated to reflect the change in matching algorithm and inability to open EPS image files on Windows machines
Originally, 'glob.glob' was being used to match the data extracted from the vflag, best_fit_param, and mass_estimate tables. When new environmental classifications were given from the output of 'void_finder,' the order of 'glob.glob' could not be preserved. The extracted data is now matched according to 'MaNGA_plate' and 'MaNGA_fiberID.' In the future, an index will be added to the master file so that the 'rotation_curve_vX_X' and 'dark_matter_mass_vX_X' can match galaxies easier. However, the output from void_finder is not directly available to these scripts. The matching method using 'MaNGA_plate' and 'MaNGA_fiberID' will have to be used until the output of void_finder can be re-ran to contain a galaxy index to simplify matching.

In addition, EPS files cannot be opened on Windows machines. Therefore, an EPS Viewer must be installed. However, most (if not all) of these software programs cannot open EPS files without the '.eps' extension. Thus, the lines saving the various images were updated to always include the matching extension specified in the 'IMAGE_FORMAT' variable.