Updated appended values for masked annuli
If all data points at a given annulus were masked, all values for that annulus were set to np.nan. This caused issues in fitting with scipy.optimize.curve_fit(). Thus, if there are no unmasked data points at an annulus, no values are appended to the data arrays for that annulus.