Created a fit_data() function
The try-except block was being executed multiple times for the positive, negative, and average rotation curve data. The creation of this function allows for better readability.

In addition, obsolete code was removed. Also, the block to plot the rotation curve data was also removed as the image file should already exist upon running this program.

Conditional statements were updated for better readability and intuitive feel.